Refactored util package
